+++ common/hlm_communication/src/builders/message_parts/mod.rs
@@ -1,4 +1,0 @@
-pub mod binary;
-pub mod parsable_part;
-pub mod string;
-pub mod type_part;