+++ common/hlm_communication/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod builders;
-pub mod certgen;
-pub mod config;
-pub mod errors;