+++ server/migrations/2026-09-12-131420-0000_create_service/down.sql
@@ -1,0 +1,3 @@
+DROP TABLE IF EXISTS service;
+DROP TABLE IF EXISTS backup_config;
+DROP TABLE IF EXISTS ip_needs_update;