+++ server/migrations/2026-09-12-131839-0000_create_ip_needs_update/down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS ip_needs_update;