+++ server/migrations/2026-09-12-132050-0000_create_user_variable/up.sql
@@ -1,0 +1,5 @@
+CREATE TABLE user_variable (
+    id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
+    id_str VARCHAR(100) NOT NULL UNIQUE,
+    value TEXT
+);