+++ server/migrations/2026-09-12-132235-0000_create_user_var_needs_update/down.sql
@@ -1,0 +1,2 @@
+DROP TABLE IF EXISTS user_variable;
+DROP TABLE IF EXISTS service;