+++ server/migrations/2026-09-12-132235-0000_create_user_var_needs_update/up.sql
@@ -1,0 +1,6 @@
+CREATE TABLE user_var_needs_update (
+    id INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
+    service_id INTEGER NOT NULL REFERENCES service(id) ON DELETE CASCADE,
+    user_variable_id INTEGER NOT NULL REFERENCES user_variable(id) ON DELETE CASCADE,
+    last_value TEXT NOT NULL
+);