+++ server/src/main.rs
@@ -1,5 +1,11 @@
 mod config;
 mod consts;
+mod context;
+mod database;
+mod logger;
+mod models;
+mod protocol;
+mod schema;
 mod server;
 mod utils;